// src/app/models/users.model.ts
export interface User {
    id: number;
    userCode: string;
    firstName: string;
    lastName: string;
    dob: string;
    email: string;
    address: string;
    city: string;
    district: string;
    town: string;
    contactFName: string;
    contactLName: string;
    gender: string;
    phone: string;
    contactPhone: string;
    password: string;
    role: string;
    image: string;
    status: string;
    file: any;
}